Create README.md
I want to prepare a platform which will be for freelancers. 70 percent is expected to address homework/assignments/essays or other related tasks for students who need assistance with class work. The 30 percent will be for other general freelancers with different tasks. I want the platform to operate in such a way that the communication can be done within the platform but no exchange of contacts is allowed. After seeing the instructions, the freelancers can bid. The bids are reviewed by the admins for the platform and a small margin added by the admins on top of freelancers bid. The payment should be paid to an escrow account and paid to the freelancer after the assignment has been approved by the client. The time zones should reflect real time in different zones. Both the clients should upload their verification details i.e ID, photo evidence and form of payment and email. The time allocated for assignment for example if deadline is at 2pm, the dashboard to both ends of the client and one doing job should display time remaining to the deadline. The users should be able to rate and review one another. There should be space where clients can sell their digital products such as e books, pieces of art, designs etc. Also in clients profiles they are allowed to display some of their works and advertise themselves. But any transaction must be done within the platform. Taskfect is the name of the platform. The users communication/information should be secure. Include grammar check and plagiarism and plagiarism checks automated. With the help of AI I want to start on this project immediately and do it myself A-Z. Let me know if it is possible.